+++ database/seeders/DatabaseSeeder.php
@@ -2,7 +2,6 @@
 
 namespace Database\Seeders;
 
-use App\Models\User;
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
 
@@ -18,12 +17,5 @@ class DatabaseSeeder extends Seeder
         $this->call(SystemPermissionsSeeder::class);
         $this->call(FirstSuperadminSeeder::class);
         $this->call(KitchenDepartmentsSeeder::class);
-
-        // User::factory(10)->create();
-
-        User::factory()->create([
-            'name' => 'Test User',
-            'email' => 'test@example.com',
-        ]);
     }
 }